NF: Move (advanced) stats to a single package
