+++ AnkiDroid/src/main/java/com/ichi2/libanki/stats/AdvancedStatistics.java
@@ -15,7 +15,7 @@
  * this program.  If not, see <http://www.gnu.org/licenses/>.                           *
  ****************************************************************************************/
 
-package com.ichi2.libanki.hooks;
+package com.ichi2.libanki.stats;
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -29,7 +29,7 @@ import com.ichi2.anki.stats.StatsMetaInfo;
 import com.ichi2.libanki.Collection;
 import com.ichi2.libanki.Consts;
 import com.ichi2.libanki.Decks;
-import com.ichi2.libanki.Stats;
+import com.ichi2.libanki.stats.Stats;
 
 import com.ichi2.utils.JSONObject;
